requirements.txt
streamlit
pandas
requests
python-binance
TA-Lib
numpy
